inference/app/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
transformers
sentencepiece
python-multipart

